Bugs and syntax fixed for website to work
Value added for author so it doesn't take value of title anymore. Book doesn't submit if form isn't fully filled. Checkbox works when ticked or unticked. Delete button fixed and syntax sorted in event listener so button works as required. When submit is clicked, a new book item is added to the list